core: add parameters to data types
This switches the representation of data types from using pi types to
encode the arguments, to using parameters and index telescopes on the
type constructor, and an argument telescope on data constructors.  When
evaluating the types of data constructors, the parameters are read off
of the type and in scope before the arguments to the data constructor.