core: use spines of elims for neutral values
This changes the representation of neutrals to use a spine of
eliminators, which is helpful to get to the head of a neutral in
constant time. This will be helpful when we are implementing
unification.  This also adds a bit more support for metavariable values,
but it is probably not working correctly.